refactor: changes edit path
